refactor: Clean-up shared state for EditPost
