+++ src/context/DataContext.js
@@ -1,7 +1,4 @@
 import { createContext, useState, useEffect } from 'react';
-import { useHistory } from 'react-router-dom';
-import { format } from 'date-fns';
-import api from '../api/posts';
 import useAxiosFetch from '../hooks/useAxiosFetch';
 
 const DataContext = createContext({});
@@ -11,9 +8,6 @@ const DataProvider = ({ children }) => {
   const [search, setSearch] = useState('');
   const [searchResults, setSearchResults] = useState([]);
 
-  const [editTitle, setEditTitle] = useState('');
-  const [editBody, setEditBody] = useState('');
-  const history = useHistory();
   const { data, fetchError, isLoading } = useAxiosFetch(
     'http://localhost:3500/posts'
   );
@@ -31,30 +25,6 @@ const DataProvider = ({ children }) => {
     setSearchResults(filteredPosts.reverse());
   }, [posts, search]);
 
-  const handleEdit = async (id) => {
-    const datetime = format(new Date(), 'MMMM dd, yyyy HH:mm:ss');
-    const editedPost = {
-      id,
-      title: editTitle,
-      body: editBody,
-      datetime: datetime,
-    };
-    try {
-      const response = await api.patch(`/posts/${id}`, editedPost);
-
-      setPosts(
-        posts.map((post) =>
-          post.id === id ? { ...response.data } : post
-        )
-      );
-      setEditTitle('');
-      setEditBody('');
-      history.push('/');
-    } catch (error) {
-      console.log(`Something went wrong: ${error.message}`);
-    }
-  };
-
   return (
     <DataContext.Provider
       value={{
@@ -65,7 +35,6 @@ const DataProvider = ({ children }) => {
         isLoading,
         posts,
         setPosts,
-        handleEdit,
       }}
     >
       {children}
